Just a zooming check in, the building position is wrong and alphas as not working yet.
git-svn-id: https://svn.code.sf.net/p/highrisedev/code@79 f146f8d7-247d-460d-a1b3-d0a612191574

+++ tiler.cpp
@@ -16,7 +16,7 @@
 
 #include "highrisedev.h"
 
-Tiler::Tiler (Texture* pTex, draw_direction direction, int x, int y, int z, int width, int height)
+Tiler::Tiler (Texture* pTex, draw_direction direction, float x, float y, float z, float width, float height)
 {
    mpTexture = pTex;
    mClipMode = true;
